+++ model_navigator/framework_api/commands/export/exporters/sm.py
@@ -1,0 +1,32 @@
+# Copyright (c) 2021-2022, NVIDIA CORPORATION. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+import fire
+import tensorflow as tf  # pytype: disable=import-error
+
+
+def get_model():
+    raise NotImplementedError("Please implement the get_model() function to reproduce the export error.")
+
+
+def export(
+    exported_model_path: str,
+):
+    model = get_model()
+    tf.keras.models.save_model(model=model, filepath=exported_model_path, overwrite=True)
+
+
+if __name__ == "__main__":
+    fire.Fire(export)